refactor: 응집도를 위한 패키지 구조 변경
